pyproject.toml
[tool.poetry]
name = "tornaracasa-address"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
SQLAlchemy = "^2.0.36"
pydantic = "^2.9.2"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
